--- model/src/main/java/com/dar/training/libraryu/model/account/AccountState.java
@@ -1,8 +1,0 @@
-package com.dar.training.libraryu.model.account;
-
-/**
- * Created by dar on 11/22/14.
- */
-public enum AccountState {
-    OPENED, ACTIVE, FROZEN, CLOSED
-}
