--- model/src/main/java/com/dar/training/libraryu/model/account/Student.java
@@ -1,23 +1,0 @@
-package com.dar.training.libraryu.model.account;
-
-/**
- * Created by dar on 11/23/14.
- */
-public class Student extends User {
-
-    private String faculty;
-
-    public Student(Long id, String name, String faculty) {
-        super.setId(id);
-        super.setName(name);
-        this.faculty = faculty;
-    }
-
-    public String getFaculty() {
-        return faculty;
-    }
-
-    public void setFaculty(String faculty) {
-        this.faculty = faculty;
-    }
-}
