--- model/src/main/java/com/dar/training/libraryu/model/account/User.java
@@ -1,26 +1,0 @@
-package com.dar.training.libraryu.model.account;
-
-/**
- * Created by dar on 11/22/14.
- */
-public class User {
-
-    private Long id;
-    private String name;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
