--- model/src/main/java/com/dar/training/libraryu/model/catalog/item/Book.java
@@ -1,17 +1,0 @@
-package com.dar.training.libraryu.model.catalog.item;
-
-/**
- * Created by dar on 11/23/14.
- */
-public class Book extends Item {
-
-    private String edition;
-
-    public String getEdition() {
-        return edition;
-    }
-
-    public void setEdition(String edition) {
-        this.edition = edition;
-    }
-}
