--- model/src/main/java/com/dar/training/libraryu/model/catalog/item/Magazine.java
@@ -1,17 +1,0 @@
-package com.dar.training.libraryu.model.catalog.item;
-
-/**
- * Created by dar on 11/23/14.
- */
-public class Magazine extends Item {
-
-    private String issue;
-
-    public String getIssue() {
-        return issue;
-    }
-
-    public void setIssue(String issue) {
-        this.issue = issue;
-    }
-}
