+++ model/src/main/java/com/darcardona/libraryu/model/account/AccountState.java
@@ -1,0 +1,8 @@
+package com.darcardona.libraryu.model.account;
+
+/**
+ * Created by dar on 11/22/14.
+ */
+public enum AccountState {
+    OPENED, ACTIVE, FROZEN, CLOSED
+}
