+++ model/src/main/java/com/darcardona/libraryu/model/catalog/Category.java
@@ -1,0 +1,36 @@
+package com.darcardona.libraryu.model.catalog;
+
+import java.io.Serializable;
+
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
+
+/**
+ * Created by dar on 11/22/14.
+ */
+public class Category implements Serializable {
+
+	private static final long serialVersionUID = 1L;
+
+	private Long id;
+
+	@NotNull
+	@Size(min = 2, max = 60)
+	private String name;
+
+	public Long getId() {
+		return id;
+	}
+
+	public void setId(Long id) {
+		this.id = id;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+}
