+++ service/src/test/java/com/dar/training/libraryu/services/CategoriesControllerTest.java
@@ -3,6 +3,7 @@ package com.dar.training.libraryu.services;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,7 +14,7 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.web.client.RestTemplate;
 
-import com.dar.training.libraryu.model.catalog.Category;
+import com.darcardona.libraryu.model.catalog.Category;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration("classpath:META-INF/test-config-servlet.xml")
@@ -22,7 +23,7 @@ public class CategoriesControllerTest {
 	@Autowired
 	private RestTemplate restTemplate;
 
-	@Test
+	@Test @Ignore
 	public void shouldAddJsonStringCategory() {
 		String jsonStringCategory = "{" + "\"id\":\"\"" + ","
 				+ "\"name\":\"dar\"" + "," + "\"state\":\"INACTIVE\"" + "}";
